+++ code/level.py
@@ -11,6 +11,7 @@ import pytmx
 from hanoi import Hanoi
 from collections import deque
 #from enemy import Enemy
+from challenge_search import *
 import random
 
 class Level:
@@ -241,7 +242,6 @@ class Level:
 
     def run(self):
         global DIFICULDADE
-
         # Executa a lógica principal do nível
         self.visible_sprites.custom_draw(self.player)
 
@@ -290,116 +290,6 @@ class Level:
         # debug(f"sorting_challenge_complete: {self.sorting_challenge_complete}", 100)
         # debug(f"Current map: {self.map_name}", 150)  # Adiciona a linha de debug para o nome do mapa
 
-class BFS:
-    def __init__(self,difficulty='easy', display_surface=None):
-        self.difficulty = difficulty
-        self.display_surface = display_surface
-        self.win = False
-        self.visited_rooms = []
-        self.rooms = {
-            'room0': ['room1_up', 'room1_down', 'room1_left', 'room1_right', 'room1_down_right'],
-            'room1_up': ['room2_up_left', 'room2_up_right'],
-            'room1_down': ['room2_left_down', 'room2_down'],
-            'room1_left': ['room2_up_left', 'room2_left'],
-            'room1_right': ['room2_right', 'room2_up_right'],
-            'room1_down_right': ['room2_down_right', 'room2_right'],
-            'room2_up_left': ['room3_up_left', 'room3_up'],
-            'room2_up_right': ['room3_up_right', 'room3_up'],
-            'room2_left_down': ['room3_down_left'],
-            'room2_down': ['room3_down_left', 'room3_down_right'],
-            'room2_right': ['room3_right_down', 'room3_right_up'],
-            'room2_down_right': ['room3_down_right'],
-            'room2_left': ['room3_left', 'room3_left_up', 'room3_left_down'],
-            'room3_up_left': [],
-            'room3_up': [],
-            'room3_up_right': [],
-            'room3_down_left': [],
-            'room3_down_right': [],
-            'room3_right_down': [],
-            'room3_right_up': [],
-            'room3_left': [],
-            'room3_left_up': [],
-            'room3_left_down': [],
-        }
-        if difficulty == 'hard':
-            self.required_path = [('room0'),
-                                  ('room1_up', 'room1_down', 'room1_left', 'room1_right',
-                                   'room1_down_right'), ('room2_up_left',
-                                'room2_up_right', 'room2_left_down', 'room2_down',
-                                'room2_right', 'room2_down_right', 'room2_left'),
-                                ('room3_up_left','room3_up','room3_up_right',
-                               'room3_down_left','room3_down_right','room3_right_down',
-                               'room3_right_left','room3_left','room3_left_up','room3_left_down')]
-        elif difficulty == 'medium':
-            self.required_path = [('room0'),
-                                  ('room1_up', 'room1_down', 'room1_left', 'room1_right',
-                                   'room1_down_right'), ('room2_up_left',
-                                'room2_up_right', 'room2_left_down', 'room2_down',
-                                'room2_right', 'room2_down_right', 'room2_left')]
-        elif difficulty == 'easy':
-            self.required_path = [('room0'),
-                                ('room1_up','room1_down', 'room1_left', 'room1_right',
-                                'room1_down_right')]
-        self.current_tuple_index = 0
-        self.current_tuple_visited = set()
-
-    def is_complete(self):
-        return self.win
-
-    def visit_room(self, room):
-        if self.win == False:
-            current_tuple = self.required_path[self.current_tuple_index]
-            if isinstance(current_tuple, tuple):
-                if (room not in current_tuple) and (room not in self.visited_rooms):
-                    print("Caminho errado")
-                    return False
-            else:
-                if room != current_tuple:
-                    print("Caminho errado")
-                    return False
-
-            if room in self.rooms:
-                if room not in self.visited_rooms:
-                    self.visited_rooms.append(room)
-                    self.current_tuple_visited.add(room)
-                    self.check_path()
-                return True
-            print("Caminho errado")
-            return False
-    def bfs(self, start_room, target_room):
-        visited = set()
-        queue = deque([start_room])
-
-        while queue:
-            current_room = queue.popleft()
-            if current_room == target_room:
-                return True
-            if current_room not in visited:
-                visited.add(current_room)
-                queue.extend(self.rooms[current_room])
-        return False
-
-    def check_path(self):
-        current_tuple = self.required_path[self.current_tuple_index]
-        if isinstance(current_tuple, tuple):
-            if all(room in self.current_tuple_visited for room in current_tuple):
-                self.current_tuple_index += 1
-                self.current_tuple_visited.clear()
-        else:
-            if current_tuple in self.current_tuple_visited:
-                self.current_tuple_index += 1
-                self.current_tuple_visited.clear()
-
-        if self.current_tuple_index == len(self.required_path):
-            print("Path completed successfully!")
-            self.win = True
-            return True
-        else:
-            print(f"Current path: {self.visited_rooms}")
-#        if self.check_level_completed():
- #           return 'next_level'
-
-
 class YSortCameraGroup(pygame.sprite.Group):
     def __init__(self):
         # Inicialização do grupo de sprites com câmera
@@ -442,3 +332,7 @@ class YSortCameraGroup(pygame.sprite.Group):
         for sprite in sorted(self.sprites(), key=lambda sprite: sprite.rect.centery):
             offset_pos = sprite.rect.topleft - self.offset
             self.display_surface.blit(sprite.image, offset_pos)
+            if hasattr(sprite, 'hitbox'):
+                hitbox_rect = sprite.hitbox.copy()
+                hitbox_rect.topleft = hitbox_rect.topleft - self.offset
+                pygame.draw.rect(self.display_surface, (255, 0, 0), hitbox_rect, 2)  # Desenha o retângulo em vermelho
