+++ code/support.py
@@ -2,6 +2,7 @@ from csv import reader
 from os import walk
 import pygame
 
+'''
 def import_csv_layout(path):
     terrain_map = []
     with open(path) as level_map:
@@ -9,15 +10,7 @@ def import_csv_layout(path):
         for row in layout:
             terrain_map.append(list(row))
     return terrain_map
-
-def import_folder(path):
-    surface_list = []
-    for _, __, img_files in walk(path):
-        for image in img_files:
-            full_path = path + '/' + image
-            image_surf = pygame.image.load(full_path).convert_alpha()
-            surface_list.append(image_surf)
-    return surface_list
+    
 
 class Button:
     def __init__(self, x, y, image, scale):
@@ -61,3 +54,12 @@ class Menu:
         def draw_text(self, text, font, text_col, x, y):
             img = font.render(text, True, text_col)
             self.screen.blit(img, (x, y))
+'''
+def import_folder(path):
+    surface_list = []
+    for _, __, img_files in walk(path):
+        for image in img_files:
+            full_path = path + '/' + image
+            image_surf = pygame.image.load(full_path).convert_alpha()
+            surface_list.append(image_surf)
+    return surface_list
